carteira/requirements.txt
Django==5.0.2
python_dateutil==2.8.2
Requests==2.31.0
